fix enumerated type in class bug
Added support for declaring enumerated types in classes by (1)
creating a getter for the enumerated type that is a type method and
that returns the enumerated type, (2) removing an assumption, in two
places in function resolution, that a type function always returns a
VarSymbol (this new type method returns a TypeSymbol directly), and
(3) removing the enumerated type from within the class (after the
getter is created) and placing it outside of the class (to normalize
the enumerated type).

Fixed a good and removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13693 3a8e244f-b0f2-452b-bcba-4c88e055c3ca