Make gasnet-segment-fast work
For _chpl_comm_broadcast_global_vars:
Add PRIMITIVE_ALLOC_GVR, which allocates space for the global vars registry.
It is called once per locale at the beginning of the generated
"_heapAllocateGlobals" function. In addition to allocating that space, it sets
a flag to indicate that it has allocated the space for this registry. This
flag is needed to ensure that the registry is the first thing allocated so its
address can known by all processes.

All changes within the compiler directory are for this primitive.

This function is meant to broadcast Chapel global variables so that all
locales reference the same location.  Since all globals live on locale
0, put copies of their wide.addr fields into the global vars registry on
locale 0, and put the address of those wide.addr fields into the registry on
other locales. Then, on each locale except 0, copy the address stored by zero
into the location stored locally. Previously all locales stored the address of
the wide.addr field, and copied the thing at that address to their own copy.
This needed to change because the wide.addr field isn't necessarily in the
remote accessible segment, so can't be remotely accessed.

For _chpl_comm_broadcast_private:
This function is meant to broadcast global constants so they don't have
to be communicated at every fork. Since their locations are not on the remote
access segment, it's not possible to broadcast them from locale 0 with
_chpl_comm_puts. Instead, allocate them into the remote access segment, copy
their values into the allocated space, fork to each locale once per constant
and read them from locale 0.

Rename the previous _chpl_malloc to _chpl_malloc1 and add a new function
_chpl_malloc2. Add a new _chpl_malloc that selects and uses one of the other
two, based on the value of an int variable controlled by the comm layer. For
comm-none and comm-gasnet with segment-everything, it uses _chpl_malloc1, and
for segment-fast and large, it uses _chpl_malloc2. The new #2 version assumes
a large block of memory has been allocated for its use, pointed to by the
meminfo struct. This large block of memory is allocated when gasnet_attach
is called.

Added a new comm-layer specific function "_chpl_comm_set_malloc_type" to pick
which malloc version to use. I added a definition of this function to
comm-none and comm-gasnet, but not comm-armci, as it seems we aren't updating
that version.

Changed _chpl_calloc into a call of _chpl_malloc and a loop to set it to zero.

Changed _chpl_free to return immediately instead of freeing anything if
_chpl_malloc2 is being used.

Make the _SET_WIDE_STRING function allocate the string into the remote
accessible segment if it's not already there.

Rearranged our runtime main function a bit so _heapAllocateGlobals gets called
before anything else that tries to allocate memory.

To build GASNet with segment-fast, add --enable-segment-fast and 
--disable-aligned-segments to third-party/gasnet/Makefile in place of
--enable-segment-everything. The disable-aligned-segments is required because
of a Linux security trick to avoid return-to-libc attacks, called address
space layout randomization.

Added a Makefile variable CHPL_GASNET_CONDUIT that is set to "udp" by default
and set to "portals" by the xt-cnl specific Makefile. Replaced instances of
"udp" with this variable.

In a gasnet-segment-everything or non-GASNet environment, I believe this
change should impact our behavior only minimally.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13701 3a8e244f-b0f2-452b-bcba-4c88e055c3ca