fixed bug regarding yieldless iterators
Made it so that the return symbol for an iterator is always added to
the vector of locals that need to be added to the class.  If the
analysis of what is live at yield points is enabled, then if we never
yield anything, nothing is assumed to be live, but we still need the
returned value in the class.

Updated a good and removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13702 3a8e244f-b0f2-452b-bcba-4c88e055c3ca