disabled instantiation limit checks in tuple module
We already don't check to see if we instantiated a function or type
too many times if that type or function is defined in ChapelBase.
With this change, we no longer check for this in ChapelTuple.  This
means that we need to be careful developing in these modules because
we could run into a situation where the compiler enters an infinite
loop due to instantiations in these modules.  I think this is the
right thing to do for now, however, given our use of instantiation to
handle tuples.  We use instantiation in the implementation of folding
in the base module, so this limit can easily be reached there too.

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13703 3a8e244f-b0f2-452b-bcba-4c88e055c3ca