Add a script to set options for configure in GASNet
This script one argument, a conduit name. It reads environment variables
CHPL_GASNET_SEGMENT and CHPL_GASNET_FORCE_ALIGNED. The first can be one of
fast, large, everything.  The second is used as a boolean and is true if set
to anything but "" or "false". Also, it reads CHPL_DEVELOPER to decide
if --enable-debug should be used.

If no environment variables it reads are set and the conduit argument is not
portals, it defaults to use only --enable-segment-everything.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13705 3a8e244f-b0f2-452b-bcba-4c88e055c3ca