Temporary hack to commConfigureOpts.pl
This utility is currently so custom to GASNet that David
put a test into it so that if CHPL_COMM was not set to
GASNet, it wouldn't print anything.  While I like this
thinking, this breaks the build of the GASNet directory
if CHPL_COMM is not set to GASNet (which we traditionally
haven't required), so commented out those lines for now
until David and I get a chance to strategize on Monday.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13710 3a8e244f-b0f2-452b-bcba-4c88e055c3ca