Removing placeholder future
Removing this future, which was a placeholder to
make sure that instantiation wasn't doing too
much work for this test.  It turns out that it
wasn't -- the tuple access method got instantiated
300-some times due to the fact that homogenous
tuples use the same implementation as heterogeneous
tuples.  Meanwhile, Steve turned off instantiation
warnings for the tuple module and added the nice
--explain-instantiation flag, which will make such
mysteries easier to investigate in the future.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13718 3a8e244f-b0f2-452b-bcba-4c88e055c3ca