More mpich case-sensitivity issues
Found another mpiCC vs. mpicc difference, so did
a quick web search and found that mpiCC is being
called mpicxx in newer versions of MPI, so took
that lead and renamed files in question using that
string instead.  (Note that this is only a
superficial renaming -- building mpich1 will
still, presumably, try to create an mpicc and an
mpiCC).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13726 3a8e244f-b0f2-452b-bcba-4c88e055c3ca