Initial support for ARMCI build
This check-in should allow linux users to build mpi
and armci and the Chapel runtime for armci without
errors.  There are several settings that are more
hacked into Makefiles than I would like -- will clean
up in subsequent check-ins, but wanted to get something
in quickly for Steve to experiment with in the meantime.

To reproduce:

1) make sure you're on a linux machine

2) in third-party do "make mpich1"

3) in third-party do "make armci"

4) making sure that CHPL_COMM is set to armci, make runtime


I haven't actually tried compiling anything to see how
well or poorly it works.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13732 3a8e244f-b0f2-452b-bcba-4c88e055c3ca