Makefile for ARMCI generated code
Added a Makefile that the compiler-generated
code relies on to link properly.  At this
point, a compile of an ARMCI code generates
a link error against _chpl_comm_broadcast_global_vars
(and several other symbols that are not ours, but
which may be defined in other MPI libraries that I
haven't specified here).

I'm going to keep working on cleaning up the
Makefiles, hoping that Steve will be able to
work on the missing broadcast routine in 
parallel.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13733 3a8e244f-b0f2-452b-bcba-4c88e055c3ca