fixed bug with remote domains/arrays
Special-cased the .locale primitive for domain and array wrappers such
that it is actually applied to the wrapped class.  This is important
because cullOverReferences treats these wrapper records specially by
removing the references in order to handle the "valid lvalue" pragma
and allow array and domain slicing, for example, to work.

Removed a future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13739 3a8e244f-b0f2-452b-bcba-4c88e055c3ca