--help support for built-in configs
In updating the CHANGES file, I noticed that we didn't have any
online help for the --maxThreads flag, and then realized that
this was because we purposefully squelch the printing of
built-in config consts/vars in our help message.  This seemed
unfortunate (particularly since the description of the -nl flag
refers to the numLocales config variable), so I enabled printing
of the internal configs.  Even better would be to support
module-specific help messages (see my feature request elsewhere)
in order to create a really clear help message for these
variables, but something's better than nothing.

I didn't expose the names of the modules that built-in configs
are defined in, but instead, list them all as being "Built-in".
This required changing code to query and set them slightly,
but isn't that different from the special casing that we
were doing to squelch their help messages before.

The one other change I had to do here was to add -nl 1 to the
execopts files for Shannon's help tests so that the numLocales
default "(configured as 1)" message that showed up for GASNet
testing would also show up in non-GASNet testing to keep the
.good files common across both test runs.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13746 3a8e244f-b0f2-452b-bcba-4c88e055c3ca