Disasbled user allocation of locales
Our locales are implemented as a class, which
meant that a user could construct new instances
of that class.  This check-in disables that
capability by throwing a conditional halt into
the constructor that is turned on after we
construct the initial numLocales locale instances.

Added a test to lock in this behavior.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13747 3a8e244f-b0f2-452b-bcba-4c88e055c3ca