added control flow checks and removed 'ordered'
Added checks to make sure that returns, yields, breaks, and continues
are not used in begins, cobegins, coforalls, foralls, syncs, and ons.
Added a note for the on-case for breaks and continues to let the
programmer know that this is a limitation and is not in the spec.

Added two open issues to the spec: (1) to discuss yields in begins,
cobegins, coforalls, and foralls, and (2) to discuss yields in ons.
Also updated the spec to make it match these checks except for breaks
and continues in ons.

Changed iterators in Schedules module to use for loops instead of
forall loops.  Also changed several tests for the same reason.

Added a test.  Recategorized a future as unimplemented feature: yield
in on statement.  This future now hits the above control flow check.

Removed 'ordered' from the implementation.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13748 3a8e244f-b0f2-452b-bcba-4c88e055c3ca