Changed semantics of .locale to return locale type
After some discussion with Steve this morning, changed
the semantics of .locale to refer the locale that the
expression is allocated on rather than the locale ID.
From the locale, the ID can always be obtained by
accessing its .id field.

My primary rationale for this change is that if the user
is working with a view of the locales other than the
built-in Locales() array, it is more helpful to get a
locale value than it is to get an ID that is fairly
entrenched in the built-in Locales() array.  It also
just seems cleaner and more intuitive to have .locale
return a locale type.

Implemented this by keeping the semantics of the
PRIMITIVE_GET_LOCALE primitive, but renaming it
as PRIMITIVE_GET_LOCALEID.  Added a built-in
function to ChapelLocale.chpl called chpl_int_to_locale()
that converts an integer to a locale type, and made
user applications of .locale get rewritten to wrap
this call around the PRIMITIVE_GET_LOCALEID that was
already being generated for them.

Rewrote a bunch of tests to refer to .locale.id instead
of .locale (could have also changed their .goods, but I
view the printing of the locale type as being fairly
internal and subject to change); also changed one
instance of "on Locales(x.locale)" into "on x.locale"
which was very pleasing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13750 3a8e244f-b0f2-452b-bcba-4c88e055c3ca