Removed outdated hand-setting of maxThreads
We used to have to set maxThreads manually to be 3 (a
superstitious value) in order to get the built-in modules
to initialize, as they were implemented using an on
clause.  Since then, local on clauses have been optimized
not to use a thread, so this manual setting of maxThreads
is no longer required.  This cleans up main() a bit.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13753 3a8e244f-b0f2-452b-bcba-4c88e055c3ca