Fixing a race condition in thread scheduling
[Paired with Steve on this change]

This check-in fixes a race condition that we had in thread 
scheduling.  Unfortunately, other races may still exist --
see below for details.

We had a bug in thread scheduling (demonstrated by Steve's
test_begin8 and test_begin9 tests) that occurred in the
following situation:

* 1 thread is in existence and idle, waiting on a condition
  variable for more work to be created

* a new task is created -- it notes that a thread exists,
  so signals that thread rather than creating a new one.

* before the thread gets scheduled after being woken up
  on its condition variable, a second task is created --
  it too notes that a thread (the same one) exists so
  signals that thread rather than creating a new one.

* the thread wakes up and starts the first task.  If the
  tasks are independent, it will execute the task then
  grab the second task;  however, if the first task blocks
  on the second task, then deadlock occurs because no
  second thread was ever created to handle the second task.

This test was caught by David Iten's 
multilocale/diten/remoteString.chpl test once our standard
modules contained an on clause, causing a thread to exist
before running the user code, and putting us into the
above situation.

Fixed this bug by having a thread woken up by a condition
variable perform a similar check as chpl_begin: signaling
a waiting thread if one exists, or creating a new thread
if it's legal to do so.  Put this code into a routine
called schedule_next_task() and put calls to it into 
chpl_begin() and chpl_begin_help().

In a more general form of the above, multiple threads are
put into the task queue and signal a smaller number of
threads, leaving some number of the threads unscheduled.
David Iten's mutilocale/diten/coforall_on.chpl was a test
that continued to deadlock after the above fix was made.

We addressed this issue by putting a while loop into
schedule_next_task that creates threads as long as there
are more tasks in the work queue and it is legal to do so.
This caused David's test to start working.  If some
threads complete their work while this while loop is
executing, this may cause unnecessary pthreads to be
spawned, but we believed that correctness on the tests
above was more important than efficiency for now.

On a run through the test system, a few other tests
failed, possibly indicating that deadlock conditions
still exist.  These were:

* parallel/begin/dinan/mvm_coforall
* parallel/taskPool/figueroa/TooManyThreads
* stress/deitz/test_10k_begins

Testing these by hand indicates that they often work;  I
did see Jim's test deadlock once in about 5 runs.  These
tests imply that there may yet be an issue when maxThreads
threads are created/exceeded.  For example, it wouldn't
surprise me if the while loop mentioned above was starving
things out in some way by holding onto the thread pool
lock for too long.  But that's just a wild guess.  More
study is needed, but Steve and I figured that it was
better to check in a step in the right direction to fix
some known problems, and continue to work on this
separately.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13763 3a8e244f-b0f2-452b-bcba-4c88e055c3ca