Bumping up artificial memory ceiling for out-of-memory test
This test checks to see that we generate an appropriate
message when a Chapel program runs out of memory.  To
make it complete in a reasonable amount of time, it
artificially lowers the virtual memory limit of the
process.  With recent changes, the compilation step
now exceeds the artifical limit, so I doubled it to
256M to give us some more breathing space in compiling
and yet still have the test run out of memory in a
short-ish amount of time.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13765 3a8e244f-b0f2-452b-bcba-4c88e055c3ca