fixed deadlock if maxThreads is exceeded
Added "waking_cnt" to the pthreads runtime code so that we keep track
of the number of threads that we've signaled to wakeup.  Changed the
conditional to guard signaling to taking into account the number of
running threads and the number of waking threads.  If signaling,
increment waking_cnt.  After waking, decrement waking_cnt when
running_cnt is incremented.  Removed while loop in schedule_next_task.

I see the maxThreads warning on several tests

  stress/deitz/test_10k_begins
  stress/deitz/test_10k_cobegins
  stress/deitz/test_10k_coforall

but this is plausible.  How should we handle this?  Ignore the warning
when testing or implement a -q flag for the runtime?  The coforall
test also seems to run out of memory when building up the task pool.
Why?

  warning: maxThreads is unbounded, but unable to create more than 805 threads
  test_10k_coforall.chpl:9: error: Out of memory allocating "instance of class _class_locals_begin_fn_1"

or

  warning: maxThreads is unbounded, but unable to create more than 1000 threads
  error: Out of memory allocating "task pool entry"

I only see this warning on the chap servers because the thread count
is high enough to have this issue.  By the way, the thread limit seems
to vary between chap05 and chap01 which is why the thread limit is
different.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13767 3a8e244f-b0f2-452b-bcba-4c88e055c3ca