Added --quiet/-q flag to compiler-generated executable
Added a --quiet/-q flag to the compiler-generated executable
that complements the --verbose/-v flag that we already had.
This sets the existing verbosity variable to 0 (default is 1,
--verbose sets it to 2).

Suppressed runtime warnings if verbosity level is set to 0.

Switched runtime check on the warning line number to test
for > 0 rather than != 0 so that line numbers of -1 will
be squashed.

Updated existing tests that look at help messages.

Added a few tests of the quiet flag: a pair that call the
chapel warning primitive manually; and a pair that set
maxThreads > 256 for GASNet.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13768 3a8e244f-b0f2-452b-bcba-4c88e055c3ca