portals conduit work for XTs
This should only change the way we act on XTs if CHPL_COMM is gasnet. In the
gasnet Makefile, compile it for target platform instead of host. Also, instead
of doing the configure directly, use the XT cross-configure script. Also, since
the regular nptl/pthreads.h header is bad on XTs, use a configure tool to
patch it, then copy it to the GASNet install directory.

In the Makefile.xt-cnl makefile, added GEN_LFLAGS and GEN_CFLAGS to use the
nptl pthreads library and the patched header.

In the broadcast private function, moved a malloc out of a loop since it only
needs to be done once.

The binary generated on an XT when using GASNet needs to be launched from a
shared filesystem (/lus/...) and can be launched as:
aprun -n<numNodes> ./a.out <options>


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13774 3a8e244f-b0f2-452b-bcba-4c88e055c3ca