Tightened up prediff script
Tightened up this prediff script to filter out more of the
filename, as the standard module that the warning is
reported in seems to vary from platform to platform.

This strangeness is similar to the foldIntParams-nomain.chpl
test that I checked in earlier today, so I'm not making
a distinct future for it.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13775 3a8e244f-b0f2-452b-bcba-4c88e055c3ca