Support for querying util scripts/environment settings
This change replaces queries of environment variables
like CHPL_THREADS and CHPL_COMM with invocations of
the util/ scripts.  This is preferable because the
util scripts choose intelligent defaults based on
the results of other util scripts (e.g., the default
value of threads will be based on the target platform).

Previously, we were making default choices of things 
like --serial and --local based only on environment 
variables and not what the util scripts would report.
This fixes that faulty logic.  It also allows Sam to
clean up his implementation of single and sync
variables by making environment settings available
to the standard modules as config params (see next
paragraph).  The combination of these two
things is what caused me to undertake such a 
significant change in the final hours before our
code freeze.

This change also makes the values of CHPL_THREADS
and CHPL_COMM available to the compiler both as
char*'s (to be referred to in the compiler source)
and as config params (to be referred to in the
standard modules):

* The char* values can be accessed using the routines
  get_CHPL_COMM and get_CHPL_THREADS in driver.cpp.
  This function is currently static because it isn't
  needed elsewhere, but can be made non-static if it
  isn't.

* The config param values are defined at the top of
  ChapelBase.chpl (after some commented out code)
  and are of type enum.  The valid values of the enums
  are given just before the config params themselves.

Ideally these config params would be defined in terms
of strings -- both to allow more flexible values to
be supported (like hyphens) and to prevent requiring
the set of all possible values from being hard-coded
into the compiler.  However, comparison of param strings
is not yet supported, so enums were chosen for simplicity.

Code has been written for the CHPL_*_TARGET and
CHPL_*_COMPILER variables to be defined using similar
config params, but since those values are (a) more likely
to include hyphens, and (b) more likely to vary from a
small set of predefined values, I left that code commented
out for now until we get param string comparison working.

At startup time, the compiler runs the util scripts,
putting their output into a temporary file, then reads
that output into a string, and files that string into
a static char* (for the compiler's use) and also into
David's config param table (to be treated like other
config params, which means they support being overridden
on the compiler command line, though I don't think we
should publicize this until it has proven useful).

Made some helper functions in order to accomplish this:

* mysystem_getresult() which wraps our mysystem() call and
  pipes the result into a temporary file, then reads 255 
  characters of that result into a string which it returns.

* runUtilScript() which wraps the previous call and takes 
  the name of a utility script (and flags) as arguments
  and returns the string generated by the utility script.

Previously, we had taken a similar action by hand in one
special case to query the name of the "make" utility that 
the compiler should use to build the generated code, so I
unified this code with the new routines by having it call
them.

The main disadvantage to all of this was that it required
changing the logic to open the temporary directory --
previously, if the user specified a saveCDir, we would
use that for all temporary files created by the compiler.
With this change, the environment variables need to be
queried before we parse command line options (to allow
the command line options to override their values in
the config param table), so we always create a temporary
directory.  Some of the logic around the temporary
directory had to change to ensure that we switched over
to the --savec directory if one was specified after the
command line arguments had been changed.

The other small disadvantage was that in order to call
the util scripts, we need to know the location of chplhome,
which the user may override on the command line using the
--chplhome flag.  Since the environment variables are
queried before the command line is parsed, we have to
use their CHPL_HOME environment variable setting in order
to run the util scripts.  My suggestion here would be
to remove the --chplhome flag from the compiler, which I
don't think is all that useful, and could prevent subtle
problems in ordering from arising.  A good task for
after the release, though.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13777 3a8e244f-b0f2-452b-bcba-4c88e055c3ca