Made isSimpleSyncBaseType() target dependent now that there is a CHPL_THREADS config param.
This, in turn, allows for significant simplification of the read/write methods
for sync and single variables.  In particular, the read?? and write?? macros
are no longer needed, except for on the MTA, where file name and line number
information is not needed.  Also, in order to work on platforms other than MTA,
it was necessary to resort to a kludge because the read?? macros needed to have
side effects, whereas the Chapel compiler assumes primitives do not have side
effects.  Since these side effects are not needed on the MTA, these primitives
can be implemented more cleanly and simply now, and the work arounds in
ChapelBase.chpl can now be removed.

Also, the names of the read??/write?? macros no longer begin with an underscore.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13780 3a8e244f-b0f2-452b-bcba-4c88e055c3ca