Updating my [too]ManyThreads tests
Adding a --maxThreads flag to manyThreads-inorder
to make it less susceptible to exhausting resources
and/or generating warnings.

Removing my tooManyThreads tests, as part of their
goal was to get the default value of maxThreads to
be unbounded (it is now) and the other part was to
get requeuable threads (which we're aware of, and
don't need tests to time out to remind us).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13783 3a8e244f-b0f2-452b-bcba-4c88e055c3ca