fixed a bug regarding running gasnet over ssh
In the change to support SEGMENT_FAST, we made the array of global variables
dynamic such that it is allocated on the heap.  In SEGMENT_EVERYTHING, we make
the assumption that the address of this array is the same on all processes, but
with the malloc, this is no longer always the case.  Added a static array that
is used in the SEGMENT_EVERYTHING case and this fixes the issue.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13793 3a8e244f-b0f2-452b-bcba-4c88e055c3ca