Support for GASNet build for Intel compilers
This change makes the GASNet build work for Intel
compilers (and more generally for non-gnu compilers)
by passing Chapel's choice of target compiler on
to the GASNet configure script.  Thus, if
CHPL_TARGET_COMPILER is set to intel, icc and icpc
will be sent on the GASNet configure script line.

In making this change, I also pushed the GASNet
installation into a compiler-specific subdirectory
so that one can have an Intel and Gnu build living
simultaneously (or even better, David can have three
builds living simultaneously for the three XT back-
end compilers).

Updated the top-level Makefile so that it now
supports a "make third-party" rule.  I'd avoided
this earlier since the release didn't have
such a directory.  Now that it does, might as
well support it at the top-level.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13796 3a8e244f-b0f2-452b-bcba-4c88e055c3ca