Moved a gnu-specific setting out of a gasnet Makefile
David pointed out this morning that I had a GNU-specific
flag embedded in the GASNet Makefile that caused XT
compilations to break (and leaving the flag out caused
them to work).  This flag is due to my use of setenv(),
which I'm still looking for a workaround for.  In this
checkin, I define a new Makefile variable,
SUPPORT_SETENV_CFLAGS that is set for the gnu and intel
compilers only to add this flag.  Longer-term, we should
figure out a cleaner way to do this (possibly by having
GASNet support the ability to launch gdb in xterms... :)

David, you looked like you had gone offline, but if you
could test this on the XTs to make sure that it
resolves your issues and doesn't break anything, that
would be great.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13799 3a8e244f-b0f2-452b-bcba-4c88e055c3ca