reimplemented 'here' as a 'private' locale
Removed special-casing of 'here' in scope resolution where it was
changed to a call of a primitive and removed the pragma "here".

Added the pragma "private" to 'here' and initialize it on every locale
when we are setting up the Locales array.

This change makes it so that simply using 'here' does not result in
communication on any locale.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13835 3a8e244f-b0f2-452b-bcba-4c88e055c3ca