improved iterator error messages + fixed gasnet regression
Specialized an error message for indexing into iterators or promotion
expressions.
test/arrays/deitz/promotion/test_scalar_promote10.good

Specialized iterator class types that appear in unresolved call
expressions if the actual is a promoted expression.
test/arrays/marybeth/test_arrayops.good

Specialized an error message for using an iterator or promoted
expression in the conditional expression of an if- or while-statement.

Removed three futures.

Added PRIMITIVE_SINGLE_UNLOCK and PRIMITIVE_SINGLE_READXX to the list
of primitives for which we dereference their actuals that are
references.  This should fix the new gasnet regression.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13842 3a8e244f-b0f2-452b-bcba-4c88e055c3ca