Fold param string comparisons
Added param versions of the six comparison operators for strings. Change the
_string_compare primitive from PRIMITIVE_UNKNOWN to PRIMITIVE_STRING_COMPARE.
Add a check in post resolution folding to catch calls to this primitive.

Add tests that I believe hit all of the possible cases of using these operators
for both params and regular vars. The unfolded case caught a bug in the
non-param version that caused for example, (assuming these weren't params!)
"hello" > "world" to be true. The problem was that the type of the
_string_compare primitive was listed as returnInfoBool, even though it returns
the int result of a strcmp call, where the sign of the result matters.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13844 3a8e244f-b0f2-452b-bcba-4c88e055c3ca