made implicit 'setter' argument a compile-time constant
Enabled folding over the 'setter' argument by cloning var-functions in
function resolution but still switching to use the value versions in
cullOverReferences.

Updated the spec.

Added FnSymbol::valueFunction so that var-functions can point to their
cloned value versions.

The 'setter' argument is not supported in iterators and we do not
clone iterators to create value versions.  I'm not sure whether we
need to make this limitation or not.  Filed a semantic future.

Removed two futures.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13854 3a8e244f-b0f2-452b-bcba-4c88e055c3ca