added setter-compilerError semantics
Added a note to the spec about setter arguments and compiler
diagnostic statements that explains why setter arguments cannot guard
a compiler diagnostic statement: "Note that when a variable function
is called in a context where the implicit setter argument is true or
false, both versions of the variable function are resolved by the
compiler. Consequently, the setter argument cannot be effectively used
to guard a compiler diagnostic statements."

Added a test that shows a compiler error that is guarded by a
conditional-statement evaluating the setter argument fires even if
that version of the function is not used.

I tried to make this work but there is an ordering difficulty since we
don't know the context of the call until we evaluate the function.  I
don't believe there is a simple solution to this problem and I think
this note is sufficient.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13856 3a8e244f-b0f2-452b-bcba-4c88e055c3ca