optimized resolution caching data structure
Reworked the data structure used during function resolution to cache
coercion wrappers, promotion wrappers, and instantiated functions.
The time spent determining whether a given function was in one of
these caches was becoming a bottleneck (time spent in the function
'checkMapCache').  In particular, the cache of instantiated functions
is very large; the other two caches are small and inconsequential.

The old data structure was a flat vector where each element contained
a function, a hash table defining how to create a new function, and
the new function (the cached FnSymbol).

The new data structure is a hash table where the keys are the old
functions and the values are vectors where each element contains a
hash table defining how to create a new function and the new function.

To check the cache, we now only need to loop over the entries in the
cache that are related to the function.  Before we needed to loop over
every entry in the cache to first determine if the entry was related
to the function begin checked.

This change improves the performance of the compiler significantly.
The following results show developer/debug compiling on chevy:

test/classes/bradc/arrayInClass/genericArrayInClass-otharrs.chpl

                         before  after  speedup
total compilation        102.6   73.6    1.39x
function resolution       37.7    8.5    4.44x
checkMapCache function    27.8    0.7   39.71x

test/arrays/deitz/jacobi.chpl

                         before  after  speedup
total compilation        9.96    8.23    1.21x
function resolution      3.51    1.92    1.83x
checkMapCache function   1.37    0.07   19.57x


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13869 3a8e244f-b0f2-452b-bcba-4c88e055c3ca