Preliminarily implemented a mechanism for adding multiple tasks to the task pool at once (though for now, chpl_begin is called for each task to be added, so there is actually no change in behavior).
Added the primitive type dtTaskList, which is a linked list of tasks to be added
to the task pool, the primitive PRIMITIVE_INIT_TASK_LIST, which simply results
in NULL being emitted during code gen and is used to initialize the linked list,
the primitive PRIMITIVE_PROCESS_TASK_LIST for adding the tasks in the task list
to the task pool.

At present, only cobegin statements use this new mechanism.  Each statement in a
cobegin statement is marked with the "cobegin" pragma.  Later, when this pragma
is encountered, a wrapper function is created and marked with the "cobegin
block" pragma.  When generating code for such wrapper functions, instead of
actually calling the wrapper function, a function called chpl_add_to_task_list()
is called instead, with the wrapper function as the parameter.  After all the
wrapper functions resulting from the cobegin statement have been added to the
task list, the function chpl_process_task_list() is called via the primitive
PRIMITIVE_PROCESS_TASK_LIST to actually put the tasks into the task pool.

Since chpl_begin is actually called to put each taskt into the task pool, the
functions chpl_add_to_task_list() and chpl_process_task_list() are implemented
in chplthreads.c, so all thread variants use this same mechanism.  This will
likely change in the near future, as each thread variant should have its own
implementation of these functions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13871 3a8e244f-b0f2-452b-bcba-4c88e055c3ca